requirements.txt
streamlit
tensorflow  # or keras if you're using Keras
pillow  # for image processing
numpy
pandas
shutil
os